Rename Model to Substore
