+++ src/index.js
@@ -1,5 +1,5 @@
 import { observable, computed } from 'mobx'
-import { Model, action, dispatch, call } from './model'
+import { Substore, action, dispatch, call } from './substore'
 import { Store } from './store'
 import { logger } from './middleware/logger'
 import { StoreContext } from './storeContext'
@@ -8,7 +8,7 @@ import { serialize } from './serialize'
 
 export {
   connect,
-  Model,
+  Substore,
   action,
   dispatch,
   call,
